+++ ace-mobile/app/src/main/kotlin/com/ace/mobile/data/local/database/entity/LocalSessionHistoryEntity.kt
@@ -1,30 +1,18 @@
+// app/src/main/kotlin/com/ace/mobile/data/local/database/entity/LocalSessionHistoryEntity.kt
 package com.ace.mobile.data.local.database.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
-/**
- * Historial de las últimas 5 sesiones completadas (FIFO).
- * Apéndice S9.
- *
- * Datos raw (sin procesar): duración, FC promedio, tipo, XP.
- * No discrimina categoría (Running, Cycling, etc. se muestran juntas).
- */
 @Entity(tableName = "local_session_history")
 data class LocalSessionHistoryEntity(
     @PrimaryKey
-    val sessionId: String,            // UUID de la sesión
-
-    val timestampStart: Long,         // Inicio
-    val timestampEnd: Long,           // Fin
-
-    val sportType: String,            // Tipo de actividad
-
-    val durationSeconds: Int,         // Duración total
-
-    val avgBpm: Double,               // FC promedio de la sesión
-
-    val totalBlocks: Int,             // Cantidad de bloques
-
-    val totalXp: Int                  // XP total ganada
+    val sessionId: String,
+    val timestampStart: Long,
+    val timestampEnd: Long,
+    val sportType: String,
+    val durationSeconds: Int,
+    val avgBpm: Double,
+    val totalBlocks: Int,
+    val totalXp: Long
 )
